relocate orphaned instances from rename plugin
