+++ plugins/hls-rename-plugin/src/Ide/Plugin/Rename.hs
@@ -26,7 +26,6 @@ import           Data.Hashable
 import           Data.List.Extra
 import qualified Data.Map                              as M
 import           Data.Maybe
-import           Data.Mod.Word
 import qualified Data.Text                             as T
 import           Development.IDE.Core.PositionMapping
 import           Development.IDE.Core.RuleTypes
@@ -47,12 +46,6 @@ import           Ide.PluginUtils
 import           Ide.Types
 import           Language.LSP.Server
 import           Language.LSP.Types
-
-instance Hashable Location
-instance Hashable Range
-instance Hashable Position
-instance Hashable UInt
-instance Hashable (Mod a) where hash n = hash (unMod n)
 
 descriptor :: PluginId -> PluginDescriptor IdeState
 descriptor pluginId = (defaultPluginDescriptor pluginId)
